Move UmbrellaStore from core → react
