+++ packages/liveblocks-core/src/client.ts
@@ -41,7 +41,6 @@ import {
   makeCreateSocketDelegateForRoom,
 } from "./room";
 import type { OptionalPromise } from "./types/OptionalPromise";
-import { UmbrellaStore } from "./umbrella-store";
 
 const MIN_THROTTLE = 16;
 const MAX_THROTTLE = 1_000;
@@ -140,7 +139,6 @@ export type EnterOptions<P extends JsonObject = DP, S extends LsonObject = DS> =
 export type PrivateClientApi<U extends BaseUserMeta, M extends BaseMetadata> = {
   readonly currentUserIdStore: Store<string | null>;
   readonly resolveMentionSuggestions: ClientOptions<U>["resolveMentionSuggestions"];
-  readonly umbrellaStore: UmbrellaStore<BaseMetadata>;
   readonly usersStore: BatchStore<U["info"] | undefined, string>;
   readonly roomsInfoStore: BatchStore<DRI | undefined, string>;
   readonly getRoomIds: () => string[];
@@ -607,8 +605,6 @@ export function createClient<U extends BaseUserMeta = DU>(
     currentUserIdStore,
   });
 
-  const umbrellaStore = new UmbrellaStore();
-
   const resolveUsers = clientOptions.resolveUsers;
   const warnIfNoResolveUsers = createDevelopmentWarning(
     () => !resolveUsers,
@@ -665,7 +661,6 @@ export function createClient<U extends BaseUserMeta = DU>(
       [kInternal]: {
         currentUserIdStore,
         resolveMentionSuggestions: clientOptions.resolveMentionSuggestions,
-        umbrellaStore,
         usersStore,
         roomsInfoStore,
         getRoomIds() {
