+++ packages/liveblocks-react/src/__tests__/umbrella-store/compareInboxNotifications.test.ts
@@ -1,4 +1,5 @@
-import type { InboxNotificationData } from "../../protocol/InboxNotifications";
+import type { InboxNotificationData } from "@liveblocks/core";
+
 import { compareInboxNotifications } from "../../umbrella-store";
 
 describe("compareInboxNotifications", () => {
