+++ packages/liveblocks-react/src/__tests__/umbrella-store/deleteComment.test.ts
@@ -1,5 +1,5 @@
 import { deleteComment } from "../../umbrella-store";
-import { createComment, createThread } from "../_dummies";
+import { createComment, createThread } from "./_dummies";
 
 describe("deleteComment", () => {
   it("should mark a comment as deleted in a thread", () => {
