+++ packages/liveblocks-react/src/comments/lib/select-notification-settings.ts
@@ -1,10 +1,8 @@
-import {
-  applyOptimisticUpdates,
-  type BaseMetadata,
-  nn,
-  type RoomNotificationSettings,
-  type UmbrellaStoreState,
-} from "@liveblocks/core";
+import type { BaseMetadata, RoomNotificationSettings } from "@liveblocks/core";
+import { nn } from "@liveblocks/core";
+
+import type { UmbrellaStoreState } from "../../umbrella-store";
+import { applyOptimisticUpdates } from "../../umbrella-store";
 
 export function selectNotificationSettings<M extends BaseMetadata>(
   roomId: string,
